fix(state): resolve CodeQL warnings for unreachable statements and cyclic imports
Use pytest.raises match parameter instead of post-block assertions to
eliminate false-positive unreachable-statement warnings from CodeQL.
Split two multi-block raise tests into separate methods.

Move resolve_store from watermark.py to state/__init__.py to break the
static import cycle between watermark, table_properties, and
metadata_table modules. The deferred imports were already safe at
runtime but CodeQL flagged them as cyclic.
